// test/integration/test/ports/engines/queries/filters/extended_relation_filters/_fixture/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Album: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly AlbumId: CodecTypes['pg/int4@1']['output'];
      readonly Title: CodecTypes['pg/text@1']['output'];
      readonly ArtistId: CodecTypes['pg/text@1']['output'];
    };
    readonly Artist: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly ArtistId: CodecTypes['pg/int4@1']['output'];
      readonly Name: CodecTypes['pg/text@1']['output'];
    };
    readonly Genre: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly GenreId: CodecTypes['pg/int4@1']['output'];
      readonly Name: CodecTypes['pg/text@1']['output'];
    };
    readonly MediaType: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly MediaTypeId: CodecTypes['pg/int4@1']['output'];
      readonly Name: CodecTypes['pg/text@1']['output'];
    };
    readonly Track: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly TrackId: CodecTypes['pg/int4@1']['output'];
      readonly Name: CodecTypes['pg/text@1']['output'];
      readonly Composer: CodecTypes['pg/text@1']['output'];
      readonly Milliseconds: CodecTypes['pg/int4@1']['output'];
      readonly Bytes: CodecTypes['pg/int4@1']['output'];
      readonly UnitPrice: CodecTypes['pg/float8@1']['output'];
      readonly AlbumId: CodecTypes['pg/text@1']['output'];
      readonly MediaTypeId: CodecTypes['pg/text@1']['output'];
      readonly GenreId: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Album: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly AlbumId: CodecTypes['pg/int4@1']['input'];
      readonly Title: CodecTypes['pg/text@1']['input'];
      readonly ArtistId: CodecTypes['pg/text@1']['input'];
    };
    readonly Artist: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly ArtistId: CodecTypes['pg/int4@1']['input'];
      readonly Name: CodecTypes['pg/text@1']['input'];
    };
    readonly Genre: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly GenreId: CodecTypes['pg/int4@1']['input'];
      readonly Name: CodecTypes['pg/text@1']['input'];
    };
    readonly MediaType: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly MediaTypeId: CodecTypes['pg/int4@1']['input'];
      readonly Name: CodecTypes['pg/text@1']['input'];
    };
    readonly Track: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly TrackId: CodecTypes['pg/int4@1']['input'];
      readonly Name: CodecTypes['pg/text@1']['input'];
      readonly Composer: CodecTypes['pg/text@1']['input'];
      readonly Milliseconds: CodecTypes['pg/int4@1']['input'];
      readonly Bytes: CodecTypes['pg/int4@1']['input'];
      readonly UnitPrice: CodecTypes['pg/float8@1']['input'];
      readonly AlbumId: CodecTypes['pg/text@1']['input'];
      readonly MediaTypeId: CodecTypes['pg/text@1']['input'];
      readonly GenreId: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly album: {
      readonly AlbumId: CodecTypes['pg/int4@1']['output'];
      readonly ArtistId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly Title: CodecTypes['pg/text@1']['output'];
    };
    readonly artist: {
      readonly ArtistId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly Name: CodecTypes['pg/text@1']['output'];
    };
    readonly genre: {
      readonly GenreId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly Name: CodecTypes['pg/text@1']['output'];
    };
    readonly mediaType: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly MediaTypeId: CodecTypes['pg/int4@1']['output'];
      readonly Name: CodecTypes['pg/text@1']['output'];
    };
    readonly track: {
      readonly AlbumId: CodecTypes['pg/text@1']['output'];
      readonly Bytes: CodecTypes['pg/int4@1']['output'];
      readonly Composer: CodecTypes['pg/text@1']['output'];
      readonly GenreId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly MediaTypeId: CodecTypes['pg/text@1']['output'];
      readonly Milliseconds: CodecTypes['pg/int4@1']['output'];
      readonly Name: CodecTypes['pg/text@1']['output'];
      readonly TrackId: CodecTypes['pg/int4@1']['output'];
      readonly UnitPrice: CodecTypes['pg/float8@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly album: {
      readonly AlbumId: CodecTypes['pg/int4@1']['input'];
      readonly ArtistId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly Title: CodecTypes['pg/text@1']['input'];
    };
    readonly artist: {
      readonly ArtistId: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly Name: CodecTypes['pg/text@1']['input'];
    };
    readonly genre: {
      readonly GenreId: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly Name: CodecTypes['pg/text@1']['input'];
    };
    readonly mediaType: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly MediaTypeId: CodecTypes['pg/int4@1']['input'];
      readonly Name: CodecTypes['pg/text@1']['input'];
    };
    readonly track: {
      readonly AlbumId: CodecTypes['pg/text@1']['input'];
      readonly Bytes: CodecTypes['pg/int4@1']['input'];
      readonly Composer: CodecTypes['pg/text@1']['input'];
      readonly GenreId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly MediaTypeId: CodecTypes['pg/text@1']['input'];
      readonly Milliseconds: CodecTypes['pg/int4@1']['input'];
      readonly Name: CodecTypes['pg/text@1']['input'];
      readonly TrackId: CodecTypes['pg/int4@1']['input'];
      readonly UnitPrice: CodecTypes['pg/float8@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;